requirements.txt
pytest==3.0.6
mock==2.0.0
pillow==4.0.0
